pyproject.toml
[project]
name = "agent-observability"
version = "0.1.0"
description = "Structured observability primitives for agent runs, tools, and LLM calls"
readme = "README.md"
authors = [
    { name = "Russell Leckenby", email = "russell.leckenby2000@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic>=2.0",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]

[project.scripts]
agent-obs-replay = "agent_observability.cli.replay:main"